Fix out-of-bounds error returning random file
Trying to return a random file from the UKWA or Archives Unleashed
would result in an out-of-bounds error because the mechanics of
that were not being well managed. E.g. we could find ourselves in
an infinite loop simply because we wanted page zero (1) and the
utility would try to reset where there was only one page to be
accessed.

There are a few other fixes in here as well along with some better
naming conventions fo variables.

Resolves #3