Fix the retrieval logic
There was a bunch of overly complex code in here. We've stripped it
back and provided more simpler checks preventing overflows and the
like.

solrMaxPages has been increased to 9999 (actual 10000) which is the
deepest the index can search without erroring on the UKWA side.

For users of moonshine this provides access to a greater range of
results. Good times!